% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_to_s3.R
\name{backup_to_s3}
\alias{backup_to_s3}
\title{Backups a list of folders into S3}
\usage{
backup_to_s3(bucket_name, folder_name, file_path = "~/.folders_to_backup.yml")
}
\arguments{
\item{bucket_name}{is the name of the bucket where the backup is stored.
If it doesn't exist, it's created automatically.}

\item{folder_name}{is the name of the root folder where the backup is stored.
If it doesn't exist, it's created automatically.}

\item{file_path}{is the path of the file that contains the folders to backup.
By default is ~/.folders_to_backup.yml. Example file in folder inst/config.}
}
\description{
This function backups a list of folders into S3.
It requires that you install the AWS CLI and configure
a profile called backuper in file ~/.aws/credentials.
}

