% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_load.R
\name{batch_load}
\alias{batch_load}
\title{Batch Load Packages}
\usage{
batch_load(pkgs)
}
\arguments{
\item{pkgs}{Is a vector containing the names of the packages.}
}
\description{
Installs packages if they don't exist in the system and then loads them.\cr
Ex:
packages <- c(
  'tidyverse',  
  'stringr',
  'DT')
batch_load(packages)
}

