% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_EDA_project.R
\name{create_EDA_project}
\alias{create_EDA_project}
\title{Create an Exploratory Data Analysis project}
\usage{
create_EDA_project(project_name)
}
\arguments{
\item{project_name}{Is the name of the EDA project.}
}
\description{
This function creates an architecture and structure for an EDA project.
}

