% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_tif.R
\name{create_plot_tif}
\alias{create_plot_tif}
\title{Save a ggplot2 plot to a tiff file}
\usage{
create_plot_tif(plot, filename)
}
\arguments{
\item{plot}{The ggplot2 plot to be saved.}

\item{filename}{The name of the file (without extension) to which the plot will be saved.}
}
\value{
None. This function is called for its side effect of saving a file.
}
\description{
This function saves a ggplot2 plot to a tiff file with a specified filename.
The file is saved with a resolution of 1200 dpi.
}
\examples{
\dontrun{
create_plot_tif(my_plot, "my_plot_filename")
}

}
