% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics.R
\name{data.clean}
\alias{data.clean}
\title{Clean up Qualtrics survey data}
\usage{
data.clean(data)
}
\arguments{
\item{data}{A data frame to be cleaned up}
}
\value{
A cleaned up data frame
}
\description{
This function takes in a data frame and cleans up the data by
removing unnecessary text from the beginning of each column name
and from the end of the first row.
}
\details{
It also renames the columns with the cleaned up names.
The resulting cleaned up data is returned.
}
\examples{
data <- data.frame(matrix(1:9, nrow = 3, ncol = 3))
cleaned_data <- data.clean(data)

}
