% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{data_clean}
\alias{data_clean}
\title{Clean up Qualtrics survey data}
\usage{
data_clean(data, includes = NULL)
}
\arguments{
\item{data}{A data frame to be cleaned up}

\item{includes}{A character vector of column names to include in the cleaned up data}
}
\value{
A cleaned up data frame
}
\description{
This function takes in a data frame and cleans up the data by
removing unnecessary text from the beginning of each column name
and from the end of the first row.
}
\details{
It also renames the columns with the cleaned up names.
The resulting cleaned up data is returned.
}
\examples{
\dontrun{
data <- data.frame(matrix(1:9, nrow = 3, ncol = 3))
cleaned_data <- data.clean(data)
}

}
