% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_tarfile.R
\name{do_tarfile}
\alias{do_tarfile}
\title{Compresses a set of files/folders into a tar.gz file}
\usage{
do_tarfile(folders, filename)
}
\arguments{
\item{folders}{Is a vector containing the paths of the
files/folders to compress. They can be absolute or relative.}

\item{filename}{Is the name for the resulting tar.gz file. This name will be
sufixed with year, month and day.}
}
\description{
Compresses a set of files/folders into a tar.gz file.
Saves output into working directory.
}

