% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contrasts.R
\name{get_contrasts}
\alias{get_contrasts}
\title{Get Contrasts}
\usage{
get_contrasts(model, vector)
}
\arguments{
\item{model}{A model object. Provide lmer models with interactions.}

\item{vector}{A vector. Provide a vector of the form ~ A | B.}
}
\value{
A data frame with the contrasts.
}
\description{
This function is used to get contrasts from a given model and vector.
}
\examples{
\dontrun{
score_model <- lmer(
    Score ~ Condition + Stage + Condition:Stage + (1 | PID),
    data = data,
    control = control
)

get_contrasts(score_model, ~ Stage | Condition)

kg_model <- lmer(
    Score ~ Condition + Process + Quiz +
        Condition:Quiz + Process:Quiz + Condition:Process +
        Condition:Process:Quiz +
        (1 | ParticipantID),
    data = data,
    control = control
)

get_contrasts(kg_model, ~ Quiz | Condition:Process)
}

}
