% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_conn.R
\name{get_db_conn}
\alias{get_db_conn}
\title{Opens JDBC Connection}
\usage{
get_db_conn(db_endpoint = db_endpoint, db_name = db_name,
  db_port = db_port, db_type = db_type, username = username,
  password = password)
}
\arguments{
\item{db_endpoint}{is the endpoint of the db instance.}

\item{db_name}{is the name of the database.}

\item{db_port}{is the db port.}

\item{db_type}{is the type of database (redshift, pgsql, sql, mysql)}

\item{username}{is an username with access to the db.}

\item{password}{is the password of the username.}
}
\description{
Opens JDBC Connection
}

