% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_likert_plot.R
\name{get_likert_plot}
\alias{get_likert_plot}
\title{This function generates a Likert plot from the given data.
It groups the data by the specified 'x, y', calculates the frequency of each group, and then calculates the ratio and marker for each group.}
\usage{
get_likert_plot(data, x, y, likert = c(1:5))
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{x}{The x in the data to be grouped by along with 'y'.}

\item{y}{The y in the data to be grouped by along with 'x'.}

\item{likert}{A vector specifying the levels for the Likert scale.
Default is c(1:5).}
}
\value{
A data frame with the grouped data
}
\description{
This function generates a Likert plot from the given data.
It groups the data by the specified 'x, y', calculates the frequency of each group, and then calculates the ratio and marker for each group.
}
\examples{
\dontrun{
get_likert_plot(data, x, y)
}

}
