% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lm_writeup.R
\name{get_lm_writeup}
\alias{get_lm_writeup}
\title{Create a writeup for the results of a linear regression model (lms and lmers).}
\usage{
get_lm_writeup(model, variable = NA, type = "statement")
}
\arguments{
\item{model}{A linear regression model (lms and lmers).}

\item{variable}{The variable to create the writeup for. If NA, the writeup
will be created for all variables.}

\item{type}{The type of writeup to create. Can be "statement" or "table".}
}
\value{
A string data frame with the writeup(s).
}
\description{
This function takes the results of a linear regression model and
creates a writeup for the results.
}
\details{
The format of the writeup is as follows:
     b = (Estimate/Slope), F(numDf, denDf) = (F), [t(numDf, denDf) = (t.ratio)], p = (p.value)
}
