% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggadd_trends.R
\name{ggadd_trends}
\alias{ggadd_trends}
\title{Add General Trends to a Model}
\usage{
ggadd_trends(model, max = 0, unit = 0, fontsize = 4)
}
\arguments{
\item{model}{A model to add general trends to.}

\item{max}{A numeric value representing the
maximum y-coordinate for the 'geom_text'
elements. Default is 0.}

\item{unit}{A numeric value representing the
unit increment for the y-coordinate of each
subsequent 'geom_text' element. Default is 0.}

\item{fontsize}{A numeric value representing the font size}
}
\value{
A list of 'geom_text' elements to be added to a ggplot.
}
\description{
This function adds general trends to a given
model by creating a list of 'geom_text'
elements for each variable in the model.
}
\details{
Each 'geom_text' element contains the variable
name, its estimate, and its p-value.
}
\examples{
\dontrun{
# Assuming 'model' is a valid model
geom_list <- add_gen_trends(model)

ggplot() + geom_list
}
}
