% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_recoder.R
\name{likert_recoder}
\alias{likert_recoder}
\title{Likert Recoder}
\usage{
likert_recoder(data, columns, levels)
}
\arguments{
\item{data}{A data frame containing the data to be recoded.}

\item{columns}{A vector of column names in 'data' that should be recoded.}

\item{levels}{A vector of levels in the Likert scale,
in the order they should be recoded.}
}
\value{
A data frame with the specified columns recoded.
}
\description{
This function recodes Likert scale responses in specified columns of a data frame.
}
\examples{
\dontrun{
pre_data <- likert_recoder(
    pre_data,
    c("AM", "ME", "DfME"),
    c("never", "some informal", "some formal", "lots of", "expert")
)
}

}
