% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_overlapping_cronjobs.R
\name{no_overlapping_cronjobs}
\alias{no_overlapping_cronjobs}
\title{No Overlapping Crong Jobs}
\usage{
no_overlapping_cronjobs()
}
\description{
Avoids overlapping cron jobs using a PID-file method.\cr
Usage: insert myrutils::no_overlapping_cronjobs() in the R code that you want
to execute via the cronjob. Place it at the top of the script/function.
}

