% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages_check.r
\name{packages_check}
\alias{packages_check}
\title{Check and install R packages}
\usage{
packages_check(packages)
}
\arguments{
\item{packages}{A character vector of package names to check and install.}
}
\description{
This function checks if the specified packages are installed and installs them if they are not.
}
\examples{
packages_check(c("dplyr", "ggplot2"))

}
