% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contrasts.R
\name{plot_contrasts}
\alias{plot_contrasts}
\title{Plot Contrasts}
\usage{
plot_contrasts(contrasts, tip_length = 0.03, fontsize = 4)
}
\arguments{
\item{contrasts}{A data frame containing the contrasts to be plotted.
It should have columns named 'estimate' and 'p.value'.}

\item{tip_length}{The length of the significance annotation lines.}

\item{fontsize}{The font size of the significance annotations. Default is 4.}
}
\value{
A ggplot object with the plotted contrasts
and significance annotations.
}
\description{
This function creates a plot of contrasts with significance annotations.
}
\examples{
\dontrun{
plot_contrasts(contrasts, max = 10, unit = 1)
}

}
