% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{plot_theme}
\alias{plot_theme}
\title{Custom theme for ggplot2 plots}
\usage{
plot_theme
}
\value{
A ggplot2 theme object.
}
\description{
This function applies a custom theme to ggplot2 plots.
The theme includes a transparent background,
and no major or minor grid lines.
}
\examples{
\dontrun{
p <- ggplot(mtcars, aes(mpg, disp)) +
    geom_point()
p + plot_theme()
}
}
