% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepData}
\alias{prepData}
\title{Prep data frame function for regression analysis}
\usage{
prepData(df_train, observable, df_test = NULL)
}
\description{
This function takes a input
'training' dataframe
and optional testing dataframe
removes nulls, converts string types to float
}
\details{
keywords trainingSample observable
good luck!
}
