% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoder.R
\name{recoder}
\alias{recoder}
\title{Recode values in a vector based on a mapping of old and new values}
\usage{
recoder(x, old_values, new_values)
}
\arguments{
\item{x}{A vector to be recoded}

\item{old_values}{A vector of old values to be replaced}

\item{new_values}{A vector of new values to replace the old values}
}
\value{
A recoded vector
}
\description{
This function takes in a vector and a mapping of old and new values, and returns a new vector with the old values replaced by the new values according to the mapping.
}
\examples{
x <- c(1, 2, 3, 4, 5)
old_values <- c(1, 2, 3)
new_values <- c("one", "two", "three")
recoded <- recoder(x, old_values, new_values)

}
