% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_packages.R
\name{reset_packages}
\alias{reset_packages}
\title{Remove all R packages except those specified}
\usage{
reset_packages(exclude = c())
}
\arguments{
\item{exclude}{A character vector of package names to keep.}
}
\value{
A message indicating the packages that have been removed.
}
\description{
Remove all R packages except those specified
}
\examples{
\dontrun{
reset_packages(exclude = c("tidyverse"))
}
}
