% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seek_and_destroy.R
\name{seek_and_destroy}
\alias{seek_and_destroy}
\title{Remove unwanted files/folders}
\usage{
seek_and_destroy(stuff_to_remove)
}
\arguments{
\item{stuff_to_remove}{Is a vector with names of folders or files to remove.
Ex: c('.DS_Store', '.wercker')}
}
\description{
Remove those unwanted files/folders that were generated automatically.
For example, '.DS_Store' or '.wercker'.\cr
As a security mechanism, only hardcoded file/folder names can be deleted.
Also, the seek and destroy operation only applies to ~/.
}

