% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_tarfile.R
\name{do_tarfile}
\alias{do_tarfile}
\title{Compresses a set of files/folders into a tar file}
\usage{
do_tarfile(paths, filename)
}
\arguments{
\item{paths}{is a vector containing the absolute path of the files/folders to compress.}

\item{filename}{is the name of the resulting file sufixed with year, month and day.}
}
\description{
This function compresses a set of files/folders into a tar file.
}

