% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tiff.R
\name{save_tiff}
\alias{save_tiff}
\title{Save rendered object as TIFF Image}
\usage{
save_tiff(object, filename = "tiff_object.tif", width = 6.67, height = 6.67)
}
\arguments{
\item{object}{The object to be saved.}

\item{filename}{The name of the file to be saved.}

\item{width}{The width of the image in inches.}

\item{height}{The height of the image in inches.}
}
\value{
NULL. The function is called for its side effect of saving an image.
}
\description{
This function saves the rendered object
as a TIFF image in the directory specified.
}
\examples{
\dontrun{
# Assuming 'plot' is a valid plot
save__tiff(plot)
}
}
